pyproject.toml
[project]
name = "peoples-court"
version = "0.1.0"
description = "A satirical AI agent that adjudicates social conflicts"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "psycopg[binary]>=3.3.2",
    "zstandard>=0.25.0",
    "requests>=2.32.3",
    "sentence-transformers>=5.2.1",
    "torch>=2.10.0",
    "einops>=0.8.2",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
    "google-genai>=1.60.0",
    "transformers>=4.48.0",
    "peft>=0.14.0",
    "datasets>=3.2.0",
    "evaluate>=0.4.3",
    "scikit-learn>=1.6.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["backend/src"]

[project.scripts]
judge = "peoples_court.cli:judge_main"
